Decode several related LD opcodes
Many of these opcodes are related in the sense that they are decoded the
same way as each other or opcodes that were already implemented. For
example, 01dstsrc is both LD r8, r8 and LD r8, [HL], which are similar
instructions, except the latter must go out to memory and costs extra
cycles. Because of this, I chose to split them into different opcodes.