Use proper read/write request functions
Using request_read/request_write eliminates the need to manually set the
bus mode. I chose to implement this in the CPU instead of the bus itself
because I plan to eliminate the bus altogether in favor of per-component
structs containing data, address, and mode pins. Regardless, this should
be easy to modify with a regex replace later if needed.